requirements_master.txt
fastapi
uvicorn
python-dotenv
pydantic
watchfiles
psycopg2-binary
google-generativeai
